+++ source/JACDAC/services/JDTestService.cpp
@@ -9,7 +9,7 @@ JDTestService::JDTestService(ManagedString serviceName, JDServiceMode m) : JDSer
 
 int JDTestService::handlePacket(JDPacket* p)
 {
-    DMESG("REC N: %d F: %d->%d N: %s M: %c%c%c", *((uint32_t*)p->data), p->device_identifier, p->service_number, this->name.toCharArray(), (this->mode == ClientService) ? "C":"", (this->mode == HostService) ? "H":"", (this->mode == BroadcastService) ? "B":"");
+    DMESG("REC N: %d F: %d->%d N: %s M: %c%c%c", *((uint32_t*)p->data), (uint32_t)p->device_identifier, p->service_number, this->name.toCharArray(), (this->mode == ClientService) ? "C":"", (this->mode == HostService) ? "H":"", (this->mode == BroadcastService) ? "B":"");
     return DEVICE_OK;
 }
 
@@ -17,6 +17,6 @@ void JDTestService::sendTestPacket(uint32_t value)
 {
     if (!this->device)
         return;
-    DMESG("SEND: %d FROM: %d", value, this->device->device_identifier);
+    DMESG("SEND: %d FROM: %d", value, (uint32_t)this->device->device_identifier);
     send((uint8_t*)&value, sizeof(uint32_t));
 }
